Rename utilities to common
