+++ common/tests/components/BaseTable.test.ts
@@ -1,9 +1,9 @@
 import { mount } from "@vue/test-utils";
 import { test } from "vitest";
 
-import BaseTable from "@/utilities/src/components/BaseTable.vue";
-import type { Pixel } from "@/utilities/src/types/pixel";
-import type { Table } from "@/utilities/src/types/table";
+import BaseTable from "@/common/src/components/BaseTable.vue";
+import type { Pixel } from "@/common/src/types/pixel";
+import type { Table } from "@/common/src/types/table";
 
 test("Table renders headers and bodies correctly.", ({ expect }) => {
   const maxHeight: Pixel = { value: "100px" };
